put <ws> in <factor>, not <number>
It makes more sense to keep number as a token. The whitespace belongs in factor, which is already a rule and just needs to allow space in the repetitions. This is also better pedantically because it makes it clear to the reader (of the book) exactly where this whitespace is taking effect.